requirements.txt
fastapi==0.114.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.9.1
pydantic-core==2.23.3
email-validator==2.2.0
python-dotenv==1.0.1
supabase-py==2.7.4